Changed Progress property in Goal to Double as discussed in meeting. Moved Optional from Controller to Service, it makes more sense for the Service to handle that.

+++ src/main/java/com/codeday/productivity/entity/Goal.java
@@ -41,7 +41,7 @@ public class Goal {
     private String isComplete;
 
     @Column(name = "progress")
-    private Integer progress;
+    private Double progress;
 
     @Column(name = "last_updated", columnDefinition = "TIMESTAMP")
     private Instant lastUpdated;
